Fix signed/unsigned comparison warning
git-svn-id: svn://scribus.net/trunk/Scribus@23499 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/plugins/import/pdf/slaoutput.cpp
@@ -3989,7 +3989,7 @@ QString SlaOutputDev::UnicodeParsedString(const std::string& s1)
 	if (s1.length() == 0)
 		return QString();
 	GBool isUnicode;
-	int i;
+	size_t i;
 	Unicode u;
 	QString result;
 	if ((s1.at(0) & 0xff) == 0xfe && (s1.length() > 1 && (s1.at(1) & 0xff) == 0xff))
@@ -4019,7 +4019,7 @@ QString SlaOutputDev::UnicodeParsedString(const std::string& s1)
 		// #15616: imagemagick may write unicode strings incorrectly in PDF
 		if (u == 0)
 			continue;
-		result += QChar( u );
+		result += QChar(u);
 	}
 	return result;
 }
